requirements.txt
asgiref==3.8.1 
django==5.1.4 
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
sqlparse==0.5.3 
typing-extensions==4.12.2
Pillow==11.1.0